Rename nuklear_sdl2_rawfb.h to nuiklear_sdl_rawfb.h
